User access rights to files and posts
Now, the user can control who can see a post or a file.
Any admin or supervisor will be able to see all posts and files, but,
for normal users, only open or allowed files and posts can be listed,
viewed, replied to, or downloaded.

SQL Scripts can be found at the end of the schema.sql file under v1.13